Moved Navbar to ModuleActivity
adding user type based xml
